refactor(post): user 도메인 분리로 인한 Post 관련 코드 수정
