+++ src/main/java/com/leets/backend/blog/post/controller/PostController.java
@@ -6,9 +6,12 @@ import com.leets.backend.blog.post.controller.dto.request.PostUpdateRequest;
 import com.leets.backend.blog.post.controller.dto.response.PostResponse;
 import com.leets.backend.blog.post.controller.dto.response.PostSummaryResponse;
 import com.leets.backend.blog.post.service.PostService;
+import com.leets.backend.blog.user.entity.User;
+import com.leets.backend.blog.login.security.UserDetailsImpl;
 import jakarta.validation.Valid;
 import org.springframework.http.HttpStatus;
 import org.springframework.http.ResponseEntity;
+import org.springframework.security.core.annotation.AuthenticationPrincipal;
 import org.springframework.web.bind.annotation.*;
 
 import java.util.List;
@@ -32,25 +35,38 @@ public class PostController {
     }
 
     @PostMapping("/posts")
-    public ResponseEntity<ApiResponse<PostResponse>> createPost(@Valid @RequestBody PostRequest postRequest) {
+    public ResponseEntity<ApiResponse<PostResponse>> createPost(
+            @Valid @RequestBody PostRequest request,
+            @AuthenticationPrincipal UserDetailsImpl userDetails) {
+
+        User user = userDetails.getUser(); // 로그인된 유저 정보 추출
+        PostResponse response = postService.createPost(request, user);
+
         return ResponseEntity.status(HttpStatus.CREATED)
-                             .body(ApiResponse.created(postService.createPost(postRequest)));
+                .body(ApiResponse.created(response));
     }
 
     @PutMapping("/posts/{postId}")
     public ResponseEntity<ApiResponse<PostResponse>> updatePost(
             @PathVariable Long postId,
-            @RequestParam Long userId,
-            @Valid @RequestBody PostUpdateRequest request) {
-        return ResponseEntity.ok(ApiResponse.ok(postService.updatePost(postId, userId, request)));
+            @Valid @RequestBody PostUpdateRequest request,
+            @AuthenticationPrincipal UserDetailsImpl userDetails) {
+
+        User user = userDetails.getUser(); // 현재 로그인된 유저
+        PostResponse response = postService.updatePost(postId, request, user);
+
+        return ResponseEntity.ok(ApiResponse.ok(response));
     }
 
     @DeleteMapping("/posts/{postId}")
-    public ResponseEntity<Void> deletePost(
+    public ResponseEntity<ApiResponse<Void>> deletePost(
             @PathVariable Long postId,
-            @RequestParam Long userId) {
-        postService.deletePost(postId, userId);
-        return ResponseEntity.noContent().build();
+            @AuthenticationPrincipal UserDetailsImpl userDetails) {
+
+        User user = userDetails.getUser();
+        postService.deletePost(postId, user);
+
+        return ResponseEntity.ok(ApiResponse.success(null));
     }
 }
