+++ src/main/java/com/leets/backend/blog/post/entity/Post.java
@@ -1,6 +1,6 @@
 package com.leets.backend.blog.post.entity;
 
-import com.leets.backend.blog.post.entity.User;
+import com.leets.backend.blog.user.entity.User;
 import jakarta.persistence.*;
 import java.time.LocalDateTime;
